pyproject.toml
[tool.poetry]
name = "lazy-tg-bot"
version = "0.1.0"
description = ""
authors = ["MiniDan21 <molchanov_danilka00@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyTelegramBotAPI = "4.26.0"
black = "^24.10.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
